Corrigindo as classes de Unidade com padrão estabelecido

+++ src/main/java/br/com/sig/msadmin/dataprovider/entity/UnidadeTable.java
@@ -19,10 +19,9 @@ import lombok.NoArgsConstructor;
 @Entity
 public class UnidadeTable {
 
-	//Todo Id deve ser informado com o tipo long, pelo motivo que o long tem um espaço maior de memoria do que o Integer
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(name="ic_tipo_unidade")
 	private String tipo;
